Make propType a byte instead of an int.
There are fewer than 127 property types, a byte is sufficient to store them.

This reduces serialized state modestly as the number of PropListItem objects, while not as large as the number of AST nodes, it is quite substantial (1-2%).

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=162235089

+++ test/com/google/javascript/jscomp/NormalizeTest.java
@@ -764,7 +764,7 @@ public final class NormalizeTest extends CompilerTestCase {
             "}"));
   }
 
-  private Set<Node> findNodesWithProperty(Node root, final int prop) {
+  private Set<Node> findNodesWithProperty(Node root, final byte prop) {
     final Set<Node> set = new HashSet<>();
     NodeTraversal.traverseEs6(
         getLastCompiler(), root, new AbstractPostOrderCallback() {
